Core Backend Systems Created
Additions:
Inventory Grid - Stores all items in the grid and handles adding and removing items

Inventory Item - Represents an instance of a GenericItem, contains a reference to all the slots it takes up, the Item it represents the parent grid of the item.

Generic Item - Generic Item that all other item types will inherit from. Currently stores item size and icon to be used for the front-end.