Added UI support for container and stackable items.
Description
---
Many changes have been made to allow for the UI / Front-end to support container and stackable item types. The most important change is that the InventoryGrid script is no longer a MonoBehaviour, this is primarily because the InventoryGrid should persist after an Item container window has been closed. Whilst this may cause some annoyances later down the line, ultimately it will be for the better.

In order to accommodate these changes, I have created a UI framework that allows for new windows to be opened that are draggable. This will make creating new functionality much easier and will improve extensibility.  The UI framework is kept in a separate repository, as it will most likely be used for future projects.

Closes #19

Additions & Changes
---
- **Stackable Item UI** - Stackable items are now supported within the UI, Stackable items now have a counter text object on their item object that represents how many items can be stacked.
- **Container Item UI** - Containers are now supported within the UI, container items are opened via the context menu, which creates a new UI window using the UI framework I created, with a newly created InventoryUIGrid as it's contents.
- **General Refactoring** - Several Miscellaneous changes have been made to various areas in the code.